# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nova Elite FC Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: team
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Players list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Register a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Player registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update player profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Player updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /trials:
    get:
      operationId: listTrials
      tags: [trials]
      summary: List upcoming trials
      responses:
        "200":
          description: Trials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trial"
    post:
      operationId: createTrial
      tags: [trials]
      summary: Create a trial slot (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrialInput"
      responses:
        "201":
          description: Trial created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trial"
  /trials/{id}/book:
    post:
      operationId: bookTrial
      tags: [trials]
      summary: Book a trial slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrialBookingInput"
      responses:
        "201":
          description: Trial booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrialBooking"
...
  /trial-bookings/{id}:
    patch:
      operationId: updateTrialBooking
      tags: [trials]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrialBookingUpdate"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrialBooking"
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update news article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: addGalleryImage
      tags: [gallery]
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List training schedules
      parameters:
        - name: team
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Schedules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingSchedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create training schedule (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSchedule"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [player, admin]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, firstName, lastName, role]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [player, admin]
        createdAt:
          type: string
    PlayerInput:
      type: object
      required: [userId, dateOfBirth, position, nationality, phone]
      properties:
        userId:
          type: integer
        dateOfBirth:
          type: string
        position:
          type: string
        nationality:
          type: string
        phone:
          type: string
        preferredFoot:
          type: string
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        guardianName:
          type: ["string", "null"]
        guardianPhone:
          type: ["string", "null"]
        passportUrl:
          type: ["string", "null"]
        govIdUrl:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        bio:
          type: ["string", "null"]
    PlayerUpdate:
      type: object
      properties:
        position:
          type: string
        phone:
          type: string
        preferredFoot:
          type: string
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        guardianName:
          type: ["string", "null"]
        guardianPhone:
          type: ["string", "null"]
        passportUrl:
          type: ["string", "null"]
        govIdUrl:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
    Player:
      type: object
      required: [id, userId, dateOfBirth, position, nationality, phone, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        dateOfBirth:
          type: string
        position:
          type: string
        nationality:
          type: string
        phone:
          type: string
        preferredFoot:
          type: string
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        guardianName:
          type: ["string", "null"]
        guardianPhone:
          type: ["string", "null"]
        passportUrl:
          type: ["string", "null"]
        govIdUrl:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
        user:
          $ref: "#/components/schemas/User"
        team:
          $ref: "#/components/schemas/Team"
        createdAt:
          type: string
    Team:
      type: object
      required: [id, name, ageGroup, coach, description]
      properties:
        id:
          type: integer
        name:
          type: string
        ageGroup:
          type: string
        coach:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        playerCount:
          type: integer
        trophies:
          type: integer
        winRate:
          type: ["number", "null"]
...
    TrialInput:
      type: object
      required: [title, date, location, ageGroup, spotsAvailable]
      properties:
        title:
          type: string
        date:
          type: string
        location:
          type: string
        ageGroup:
          type: string
        spotsAvailable:
          type: integer
        description:
          type: ["string", "null"]
    Trial:
      type: object
      required: [id, title, date, location, ageGroup, spotsAvailable, spotsBooked]
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        location:
          type: string
        ageGroup:
          type: string
        spotsAvailable:
          type: integer
        spotsBooked:
          type: integer
        description:
          type: ["string", "null"]
        status:
          type: string
    TrialBookingInput:
      type: object
      required: [playerName, email, phone, dateOfBirth, position]
      properties:
        playerName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        position:
          type: string
        notes:
          type: ["string", "null"]
    TrialBookingUpdate:
      type: object
      properties:
        status:
          type: string
    TrialBooking:
      type: object
      required: [id, trialId, playerName, email, phone, dateOfBirth, position, status]
      properties:
        id:
          type: integer
        trialId:
          type: integer
        playerName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        position:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    NewsInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
    NewsUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
    NewsArticle:
      type: object
      required: [id, title, content, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        createdAt:
          type: string
    GalleryImageInput:
      type: object
      required: [imageUrl, title, category]
      properties:
        imageUrl:
          type: string
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
    GalleryImage:
      type: object
      required: [id, imageUrl, title, category]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ScheduleInput:
      type: object
      required: [teamId, dayOfWeek, startTime, endTime, location, type]
      properties:
        teamId:
          type: integer
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        type:
          type: string
        notes:
          type: ["string", "null"]
    TrainingSchedule:
      type: object
      required: [id, teamId, dayOfWeek, startTime, endTime, location, type]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        type:
          type: string
        notes:
          type: ["string", "null"]
        team:
          $ref: "#/components/schemas/Team"
...