# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nova Elite FC Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleInput {
  teamId: number;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  location: string;
  type: string;
  /** @nullable */
  notes?: string | null;
}